requirements.txt
alembic==1.13.2
annotated-types==0.7.0
anyio==4.4.0
black==24.8.0
blinker==1.8.2
cfgv==3.4.0
click==8.1.7
distlib==0.3.8
fastapi==0.114.0
filelock==3.16.0
flake8==7.1.1
Flask==3.0.3
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
greenlet==3.0.3
h11==0.14.0
identify==2.6.0
idna==3.8
itsdangerous==2.2.0
Jinja2==3.1.4
Mako==1.3.5
MarkupSafe==2.1.5
mccabe==0.7.0
mypy-extensions==1.0.0
nodeenv==1.9.1
packaging==24.1
pathspec==0.12.1
platformdirs==4.3.2
pre-commit==3.8.0
psycopg==3.2.1
pycodestyle==2.12.1
pydantic==2.9.1
pydantic_core==2.23.3
pyflakes==3.2.0
python-dotenv==1.0.1
PyYAML==6.0.2
ruff==0.6.4
sniffio==1.3.1
SQLAlchemy==2.0.34
starlette==0.38.5
typing_extensions==4.12.2
uvicorn==0.30.6
virtualenv==20.26.4
Werkzeug==3.0.4
